// apps/frontend/src/lib/types/iex-cloud.ts
export interface KeyStats {
  marketCap: number
  peRatio: number
  forwardPE?: number
  eps?: number
  dividendYield?: number
  beta: number
  week52High: number
  week52Low: number
  profitMargin?: number
  revenuePerShare?: number
  priceToBook: number
  priceToSales: number
  sharesOutstanding?: number
  avgTotalVolume?: number
  avgVolume?: number
  ttmEPS?: number
  ttmDividendRate?: number
  dividendYieldPercentage?: number
  evToEbitda?: number
}